refactor: remove duplicated code and dead code across codebase
- Extract shared test helpers to tests/common/mod.rs, eliminating
  duplication across 4 test files
- Remove 7 duplicated nested get_code helper functions in LSP
  diagnostics tests, keeping the single module-level definition
- Remove unused timeout fields from all 11 price source structs

Net: ~160 lines removed.
